Add Dockerfile.kerneldoc-sarasa-focal
Ubuntu focal's pango is version 1.44.
While it dropped support of bit map fonts including Type 1 and
pfb, it does not ignore those fonts suggested by fontconfig.
This causes some texts in SVG files missing in converted PDF.

As a workaround, this Dockerfile moves such fonts to a local
directory which is not in the fontconfig path.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>